+++ modules/actor/actor-sheet.js
@@ -4,6 +4,8 @@ import { HMRollMgr } from '../mgr/rollmgr.js';
 import { HMTABLES, MODULE_ID } from '../sys/constants.js';
 
 export class HMActorSheet extends ActorSheet {
+    visibleItemId = {};
+
     /** @override */
     get template() {
         const path = 'systems/hackmaster5e/templates/actor';
@@ -20,78 +22,57 @@ export class HMActorSheet extends ActorSheet {
         return data;
     }
 
-    // TODO: This function is a mess and needs a refactor.
     _prepareBaseItems(sheetData) {
         const actorData = sheetData.actor;
-        const uskills = [];
-        const skills = [];
-        const langs = [];
-        const wounds = [];
-        const gear = {
-            'weapons': [],
-            'armors':  [],
-            'items':   [],
-        };
-        const spells = [];
-        const armors = {
-            'owned':    [],
-            'carried':  [],
-            'equipped': [],
-        };
-        const weapons = {
-            'owned':    [],
-            'carried':  [],
-            'equipped': [],
-            'innate':   [],
-        };
 
-        for (const i of sheetData.items) {
-            if (i.type === 'skill') {
-                if (i.system.language) { langs.push(i); } else {
-                    if (actorData.type === 'character') {
-                        i.system.universal ? uskills.push(i) : skills.push(i);
-                    } else { skills.push(i); }
-                }
-            } else
-            if (i.type === 'armor') {
-                gear.armors.push(i);
-                const state = HMTABLES.itemstate[(i.system.state)];
-                armors[state].push(i);
-            } else
-            if (i.type === 'weapon') {
-                gear.weapons.push(i);
-                const state = HMTABLES.itemstate[(i.system.state)];
-                weapons[state].push(i);
-            } else
-            if (i.type === 'wound')  { wounds.push(i);     } else
-            if (i.type === 'item')   { gear.items.push(i); } else
-            if (i.type === 'spell')  { spells.push(i);     }
-        }
+        const {spell} = actorData.itemTypes;
+        const gear = {weapons: [], armors: [], items: []};
+        const armors = {owned: [], carried: [], equipped: []};
+        const weapons = {owned: [], carried: [], equipped: [], innate: []};
+        const skills = {uskills: [], oskills: [], iskills: []};
 
-        // Sort
+        actorData.itemTypes.skill.forEach((i) => {
+            if (i.system.language) {
+                skills.iskills.push(i);
+                return;
+            }
+
+            if (actorData.type !== 'character' || !i.system.universal) {
+                skills.oskills.push(i);
+                return;
+            }
+
+            skills.uskills.push(i);
+        });
+
+        actorData.itemTypes.armor.forEach((i) => {
+            gear.armors.push(i);
+            const state = HMTABLES.itemstate[(i.system.state)];
+            armors[state].push(i);
+        });
+
+        actorData.itemTypes.weapon.forEach((i) => {
+            gear.weapons.push(i);
+            const state = HMTABLES.itemstate[(i.system.state)];
+            weapons[state].push(i);
+        });
+
+        gear.items = actorData.itemTypes.item;
+
         function skillsort(a, b) {
             return `${game.i18n.localize(a.name)} ${a.system.specialty.value || ''}`
                  > `${game.i18n.localize(b.name)} ${b.system.specialty.value || ''}` ? 1 : -1;
         }
 
-        skills.sort(skillsort);
-        langs.sort(skillsort);
+        Object.keys(skills).forEach((skillType) => skills[skillType].sort(skillsort));
 
-        if (actorData.type === 'character') {
-            uskills.sort(skillsort);
-            actorData.skills = {skills, uskills, langs};
-        } else {
-            actorData.skills = {skills, langs};
-        }
-
-        actorData.wounds = wounds;
+        actorData.skills = skills;
         actorData.armors = armors;
+        actorData.weapons = weapons;
         actorData.gear = gear;
-        actorData.spells = spells.sort(
+        actorData.spells = spell.sort(
             (a, b) => Number(a.system.lidx) - Number(b.system.lidx) || a.name.localeCompare(b.name),
         );
-
-        actorData.weapons = weapons;
 
         const slevels = [];
             for (let i=0; i < actorData.spells.length; i++) {
@@ -102,31 +83,7 @@ export class HMActorSheet extends ActorSheet {
         actorData.slevels = slevels.sort();
 
         // If sheet has only one spell level, the controls are locked.
-        if (slevels.length == 1) {
-            actorData.system.cslevel = actorData.slevels[0];
-        }
-    }
-
-    async _prepareCharacterItems(sheetData) {
-        const actorData = sheetData.actor;
-
-        // Initialize containers.
-        const profs = [];
-        let race = null;
-        let cclass = null;
-
-        // Iterate through items, allocating to containers
-        for (const i of sheetData.items) {
-            i.img = i.img || DEFAULT_TOKEN;
-            if (i.type === 'cclass')      { cclass = i;    } else
-            if (i.type === 'proficiency') { profs.push(i); } else
-            if (i.type === 'race')        { race = i;      }
-        }
-
-        // Assign
-        actorData.cclass = cclass;
-        actorData.profs = profs;
-        actorData.race = race;
+        if (slevels.length === 1) actorData.system.cslevel = actorData.slevels[0];
     }
 
     /** @override */
@@ -225,9 +182,14 @@ export class HMActorSheet extends ActorSheet {
 
     async _onToggle(ev) {
         ev.preventDefault();
+        const cId = this._getItemId(ev);
+
+        const tState = !this.visibleItemId[cId];
+        this.visibleItemId[cId] = tState;
+
         const element = ev.currentTarget;
-        const target  = $(element).parent().find("[toggle]");
-        $(target).toggleClass("hide");
+        const target  = $(element).parent().find('[toggle]');
+        $(target).toggleClass('hide');
     }
 
     // Toggle between an item being equipped, carried, or stored.