+++ modules/actor/character-actor-sheet.js
@@ -1,13 +1,22 @@
 import { HMActorSheet } from './actor-sheet.js';
 import { HMTABLES } from '../sys/constants.js';
 
+function prepareCharacterItems(sheetData) {
+    const {actor} = sheetData;
+    const {itemTypes} = actor;
+
+    // Assign
+    actor.race = itemTypes.race?.[0];
+    actor.cclass = itemTypes.cclass?.[0];
+}
+
 export class HMCharacterActorSheet extends HMActorSheet {
     /** @override */
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
             classes: ['hackmaster', 'sheet', 'actor'],
             width: 835,
-            height: 941,
+            height: 930,
             tabs: [
                 { navSelector: '.sheet-tabs', contentSelector: '.sheet-body', initial: 'skills' },
                 { navSelector: '.bio-tabs',   contentSelector: '.bio-body',   initial: 'wealth' },
@@ -19,7 +28,7 @@ export class HMCharacterActorSheet extends HMActorSheet {
     getData() {
         const data = super.getData();
 
-        this._prepareCharacterItems(data);
+        prepareCharacterItems(data);
         return data;
     }